refactor: address clang-tidy warnings

+++ src/graphics/arcball.h
@@ -17,8 +17,8 @@ namespace qem::arcball {
  * \see docs/arcball.pdf for a detailed description of the arcball interface.
  */
 std::optional<const std::pair<const glm::vec3, const float>> GetRotation(
-    const glm::dvec2& cursor_position_start,
-    const glm::dvec2& cursor_position_end,
+    const glm::vec2& cursor_position_start,
+    const glm::vec2& cursor_position_end,
     const std::pair<const int, const int>& window_dimensions);
 
 }  // namespace qem::arcball
